finsaferag/finsaferag/requirements.txt
absl-py==2.3.1
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiolimiter==1.2.1
aiosignal==1.4.0
aiosqlite==0.21.0
altair==5.5.0
annotated-doc
annotated-types
anyio
appdirs==1.4.4
attrs==25.4.0
banks==2.2.0
beautifulsoup4==4.14.2
bert-score==0.3.13
blinker==1.9.0
blis==1.3.3
bm25s==0.1.10
bs4==0.0.2
cachetools==6.2.2
catalogue==2.0.10
certifi
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cloudpathlib==0.23.0
cohere==5.20.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.3
cryptography==44.0.3
cycler==0.12.1
cymem==2.0.13
dashscope==1.20.0
dataclasses-json==0.6.7
datasets==3.1.0
deepeval==0.21.26
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
dnspython
docstring_parser==0.16
docx2txt==0.9
email-validator
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
evaluate==0.4.6
exceptiongroup
execnet==2.1.2
faiss-cpu==1.13.0
fastapi
fastapi-cli
fastavro==1.12.1
filelock==3.20.0
filetype==1.2.0
fire==0.7.1
FlagEmbedding==1.2.10
flwr==1.23.0
flwr-datasets==0.5.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2024.9.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.4
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-genai==1.52.0
google-generativeai==0.5.4
googleapis-common-protos==1.72.0
greenlet==3.2.4
griffe==1.15.0
grpcio==1.76.0
grpcio-health-checking==1.62.3
grpcio-status==1.62.3
h11
h2
hf-xet==1.2.0
hpack
httpcore
httplib2==0.31.0
httptools
httpx
httpx-sse==0.4.0
huggingface-hub==0.23.5
hyperframe
idna
importlib_metadata==7.0.2
iniconfig==2.3.0
instructor==1.5.2
iterators==0.0.2
Jinja2
jiter==0.5.0
jiwer==3.1.0
joblib==1.5.2
json5==0.12.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain==0.2.7
langchain-community==0.2.7
langchain-core==0.2.43
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langsmith==0.1.147
lazy_loader==0.4
litellm==1.48.10
llama-cloud==0.1.19
llama-cloud-services==0.6.21
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.7
llama-index-embeddings-huggingface==0.3.1
llama-index-embeddings-langchain==0.2.1
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.3.1
llama-index-instrumentation==0.4.2
llama-index-legacy==0.9.48
llama-index-llms-dashscope==0.2.1
llama-index-llms-huggingface==0.3.1
llama-index-llms-ollama==0.3.0
llama-index-llms-openai==0.2.16
llama-index-multi-modal-llms-openai==0.2.3
llama-index-postprocessor-cohere-rerank==0.2.1
llama-index-postprocessor-colbert-rerank==0.2.1
llama-index-postprocessor-flag-embedding-reranker==0.2.0
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-index-retrievers-bm25==0.3.0
llama-index-vector-stores-faiss==0.2.1
llama-index-workflows==2.11.5
llama-parse==0.6.21
loguru==0.7.3
lxml==6.0.2
markdown-it-py
MarkupSafe
marshmallow==3.26.1
matplotlib==3.10.7
mauve-text==0.4.0
mdurl
minijinja==2.12.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.16
murmurhash==1.0.15
mypy_extensions==1.1.0
narwhals==2.12.0
nest-asyncio==1.6.0
networkx==3.6
nltk==3.9.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
ollama==0.6.1
openai==1.109.1
orjson==3.11.4
packaging==24.2
pandas==2.2.2
pathspec==0.12.1
phonenumbers==9.0.19
pillow==10.4.0
platformdirs==4.5.0
plotly==6.5.0
pluggy==1.6.0
polars==1.35.2
polars-runtime-32==1.35.2
portalocker==3.2.0
preshed==3.0.12
presidio_analyzer==2.2.360
presidio_anonymizer==2.2.360
prettytable==3.17.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==4.25.1
psutil==7.1.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core
pydeck==0.9.1
Pygments
PyJWT==2.8.0
PyMuPDF==1.26.6
pynvml==13.0.1
pyparsing==3.2.5
pypdf==4.3.1
PySocks==1.7.1
PyStemmer==2.2.0.3
pytest==9.0.1
pytest-repeat==0.9.4
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv
python-multipart
pytz==2025.2
PyYAML
ragas==0.3.2
RapidFuzz==3.14.3
ray==2.31.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-file==3.0.1
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit
rouge-score==0.0.4
rpds-py==0.29.0
rsa==4.9.1
sacrebleu==2.5.1
safetensors==0.7.0
scikit-learn==1.3.2
scipy==1.16.3
seaborn==0.13.2
sentence-transformers==2.7.0
sentencepiece==0.2.1
sentry-sdk==2.46.0
seqeval==1.2.2
shellingham
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
sniffio
soupsieve==2.8
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.44
srsly==2.5.2
starlette
streamlit==1.51.0
streamlit-card==1.0.2
streamlit-option-menu==0.4.0
striprtf==0.0.26
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
tenseal==0.3.16
termcolor==3.2.0
text-generation==0.7.0
thinc==8.3.10
threadpoolctl==3.6.0
tiktoken==0.12.0
tldextract==5.3.0
tokenizers==0.19.1
toml==0.10.2
tomli
tomli_w==1.2.0
torch==2.2.0
torchaudio==2.2.0
torchdata==0.7.1
torchtext==0.17.0
torchvision==0.17.0
tornado==6.5.2
tqdm==4.66.4
transformers==4.42.2
triton==2.2.0
typer==0.12.5
typer-slim==0.20.0
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection
typing_extensions
tzdata==2025.2
ujson==5.11.0
uptrain==0.7.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn
uvloop
validators==0.35.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles
wcwidth==0.2.14
weasel==0.4.3
websocket-client==1.9.0
websockets
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zhipuai==2.1.5.20250825
zipp==3.23.0
